+++ changelog.md
@@ -1,3 +1,10 @@
+# 1.1.0
+- Handle ESC key (close the command palette)
+- Fix not being able to input the period character
+- Remove close button (if on PC)
+- Add a helpful tip in the command palette :)
+- Add `gd stats` command (shows your statistics)
+
 # 1.0.1
 - Mark Blur API as <cg>optional</c>
 - Prevent the command palette from being summoned inside a level